cornershop-test/requirements.txt
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
Django==2.0.2
djangorestframework==3.9.1
docutils==0.14
factory-boy==2.11.1
Faker==1.0.2
gunicorn==19.9.0
idna==2.8
isort==4.3.9
itypes==1.1.0
psycopg2-binary==2.8.4
pymongo==3.7.2
python-dateutil==2.8.0
python-dotenv==0.10.1
python-json-logger==0.1.10
python-slugify==2.0.1
pytz==2018.9
requests==2.21.0
slackclient==1.0.6
celery==4.0.2
